feat: Dealer와 Player의 공통 부분을 모아둔 Participant 추가
